AMB-221 Disables 'Try this API'
Adds custom Specification Extension property
`x-spec-publication.try-this-api.disabled`
which allows toggling visibility of the
'Try this API' button in the specification's
page on NHSD website.

Setting this property to true removes the
'Try this API' button; setting it to `false`
or removing it altogether restores the button.

Actually implemented as ADZ-140 (see for
more details and the link to documentation)
but originally raised as AMB-221.